Fix various typos
I took a sweep through the sources and text files with a spell checker.

I fixed several, but due to the nature of the sources there are many
that I can't fix easily without risking save compatibility.

The heavy use of `global` and scoped imports makes refactoring the source more difficult.
I'll need that resolved before I can fix the spelling of variables and functions.